Fix mac shared crash on game load
This is a core game problem that was replicated in ACE through the
'deferred_lighting.shader` file - there was a missing version and the
fragment shader was still using `#version 120`.

I've upgraded the shader to use `#version 410` syntax - I've seen this
in other files in Stonehearth so hoping not breaking minimum spec. I've
tested this working on my M1 running macOS Monterey 12.3.