small fix to promote script in campaigns
If a player sent to the bunnies someone that didn't had geomancer as a job option, the game would error out when trying to promote to that. E.g. an orc citizen only has worker and footman as job options

This will add the request job as one of the possible options to the limited job tree of that citizen, avoiding the error.