Tweaks to combat damage calculation
- There's now a "muscle_multiplier" weapon data parameter that defines how much will muscle affect the final damage.
Crossbows and Cleric Tomes will now have 0 influence.
Bows will have 80% instead of 100%
Offhand weapons will have 60% instead of 100%
Two-handed weapons will have 120%

Additionally, when damage is less than 1 it will now take into account how close it is to 1 for defining if it will inflict 1 or 0
For example, if final damage calculation is 0.85, that means there's 85% of chance for damage to be 1 and 15% of chance for damage to be 0.